feat(plugin): format byte size into appropriate unit
Used the human read crate to format it.
Also convert block size in blocks to a size.
The doc mention that the size is in 512 blocks, is this really always true? Even
if the device is 4k?

+++ control-plane/stor-port/src/types/v0/transport/blockdevice.rs
@@ -92,7 +92,7 @@ impl From<BlockDevice> for models::BlockDevice {
             src.filesystem,
             src.model,
             src.partition,
-            src.size as i64,
+            src.size,
         )
     }
 }
